[CL-2362] Move tooltip copy
