+++ front/app/containers/Admin/projects/project/general/components/TopicInputsTooltipExtraCopy.tsx
@@ -5,7 +5,7 @@ import Link from 'utils/cl-router/Link';
 
 // i18n
 import { FormattedMessage } from 'utils/cl-intl';
-import messages from './messages';
+import messages from '../messages';
 
 const TopicInputsCopyProvider = () => {
   return (
