+++ front/app/modules/commercial/custom_topics/admin/components/messages.ts
@@ -46,14 +46,7 @@ export default defineMessages({
     id: 'app.containers.AdminPage.SettingsPage.AllowedInputTopics.topicsTermPlaceholder',
     defaultMessage: 'tags',
   },
-  topicInputsTooltipExtraCopy: {
-    id: 'app.containers.AdminPage.SettingsPage.AllowedInputTopics.topicInputsTooltipExtraCopy',
-    defaultMessage: 'Tags can be configured {topicManagerLink}.',
-  },
-  topicInputsTooltipLink: {
-    id: 'app.containers.AdminPage.SettingsPage.AllowedInputTopics.topicInputsTooltipLink',
-    defaultMessage: 'here',
-  },
+
   tagIsLinkedToStaticPage: {
     id: 'app.containers.AdminPage.ProjectEdit.tagIsLinkedToStaticPage',
     defaultMessage:
